Adding data for a variety of timeframes
Bunch of data files. The time frame (time represented by each candle) is the 2nd block set off by a hyphen. The total amount of time represented in the data set is the 3rd block in the filename. Included the date because I will need to update these periodically as time goes on. After 10 more trading days pass, I will collect the info again and add a new Excel file. We can concatenate them into one main dataframe/excel file and track different timeframes for longer periods (the max allowed by TDA is 10 days on anything less than a daily candle.

An unfortunate note: they do not have hourly candles as an option, but the would be really nice to use for analysis. We will have to process the available timeframes into other timeframes if we hope to include them in our analysis. Shouldn't be too hard with dataframe comprehension tools though.